+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes will be documented in this file.
 
 --- 
 
+## [1.1.0](https://github.com/apprack/apprack-ios-sdk/releases/tag/1.0.1) (2018-11-17)
+
+* Improved support for iOS 13 Dark Mode.
+* APRSettingControllor: 
+  * Request a refresh of the screen.
+  * Option to remove custom views.
+
 ## [1.0.1](https://github.com/apprack/apprack-ios-sdk/releases/tag/1.0.1) (2018-11-17)
 
 * Improved error handling.